+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- expects and parseJson behavior now properly parse content types. #22
 - Refactored Boostrap #19 to not use event emitter to allow wrapping
 
 ## [0.3.1] - 2017-04-01